app/backend/requirements.txt
fastapi
uvicorn
torch
transformers
sentence-transformers
pandas
numpy
faiss-cpu
kagglehub
